requirements.txt
python-telegram-bot>=20.0
notion-client>=0.13.0
python-dotenv>=1.0.0